Predict attribute from examples
  ref #15
